Update patch set 1
The concern of the directory being deleted while the search context is open is handled thus:

Before the directory can be deleted, all the objects in the directory will first have to be unlinked. That will cause the removeObjectCallback to be called for each object in the directory. That will cause SearchAdvance to be called for each object which means the search context will be positioned so that sc->nextReturn is NULL. 

Once sc->nextReturn is NULL, the search context is "zombied" and it no longer looks at the directory or any other objects.

Patch-set: 1