Update patch set 1
Patch Set 1:

We are adding a dependency to the API when we do not need to - thats my issue with this. 

This being US specific and the dependency on country code is how the driver manages it at this time. At some point later, this could be changed to work on other countries in the driver.

From an extensibility perspective (for other drivers that might do this differently), it is good if the APIs are independent.

Patch-set: 1