Update patch set 3
Patch Set 3:

It would be more flexible if the decision of whether to use an LD or the ttyGS could be made at runtime by userspace, rather than in the board file at compile-time.

Is there a reason the ldisc needs to be a separate function from acm?  Is it possible for the existing acm function to register as an LD, and simultaneously expose a ttyGS to userspace?

Patch-set: 3