Update patch set 4
Patch Set 4:

Rob,

I'm working on some supporting documentation to describe the problem we're trying so solve in more detail, but in short, this is intended as the api and first look at what we're building.  Basically we're trying to create a single unified "memory manager" for all the soc's we encounter that allows us to do zero copy gfx, video playback, camera preview etc and any mix of those.  On pretty much every program, we've written or inherited a new solution (sometimes several hybrid solutions we had to glue together) to solve this problem.  

I'm not a GEM expert by any means, but from what I can tell it doesn't solve a couple of problems we need to.  First off I don't think it has any mechanism for securely sharing handles between userspace processes.  That's essential on Android.  We pass buffers around a lot.  Secondly, and probably more importantly, from what I can tell GEM mostly addresses sharing needs between the gpu and cpu.  We care about sharing between the gpu, cpu, isp block, video encoder and decoders etc.  On different SOC's those have all different memory requirements -- in some cases separate iommu's, one iommu, special purpose mmus, a simple gart, direct physical memory access, etc.  For testing and clarity I've implemented two REALLY simple heaps and a mapper, but the intention is for the sets of heaps and mappers to grow to cover all these cases.  

As you've pointed out, one of the things we'll be adding is a heap for managing large chunks of physically contiguous memory.  Additionally, we'll be doing a lot of work to manage the complex requirements of the tiler on omap4 as a mapper.  

The intention of posting what I have so far was to start eliciting feedback, so please do let me know what you think.

Thanks, 
Rebecca

Patch-set: 4