Update patch set 5
Patch Set 5:

ok, some of the missing bits make more sense when I realize it is just intended to be a sort of preview of the API.  I didn't understand that part initially.

re: securely sharing buffers:

GEM (or rather DRM) provides an authentication mechanism, so a client would need to be connected/authenticated with the window manager.. xorg in generic linux case, but I guess that would be surface flinger in the android case.  

Basic way it works is that client process gets a token from the kernel (drmGetMagic()), passes it to display server over whatever IPC mechanism, which then passes it back to the kernel (drmAuthMagic()).  This unlocks for the client all the ioctls that have the DRM_AUTH flag set.

This seems well designed for the use cases I can think of, but would of course be interested if it seems to be missing something from your PoV.

----

And yes, you are right that GEM is somewhat oriented towards GPU<->CPU, and DRM driver framework tends to lump together a lot of different useful pieces (mode setting, buffers, hotplug, etc)..  but in the desktop world it solves the problem of sharing buffers between client process and display server.  It may need some extension to support cameras and other sort of things.

In general I prefer the idea of extending an existing solution rather than introduce a new one which will (I'm sure) lead to a bigger fork between generic linux and android kernel.  By that I mean that we'd have to implement display driver, multimedia accelerators, etc, differently between android and generic kernels.

There is a lot of interest in the SoC world (at ELC right now.. btw if you have time to come over for the BoF sessions this evening at 6:30) about having a good common solution for passing buffers around.  Whether the eventual solution is DRM/GEM, or ION, or something else, it would be nice to have a common solution between upstream linux and android kernel.    That is my biggest concern.

Patch-set: 5