Update patch set 1
Patch Set 1:

The new UHID driver seems a good replacement of this CL. But I am so busy and this is not a simple swap of /dev/bthid with /dev/uhid. Bluedroid stack has to re-implement the CO functions. Filed a bug to track this: b/6963714 In the meantime.

Patch-set: 1
Reviewer: Matthew Xie <1013388@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0