Update patch set 3
Patch Set 3:

As of 3.10 I believe we shouldn't need at least the CPUFREQ_GOV_START check, since the governor callbacks won't be called with offline cpus in the policy.  We may need a check and a get/put_online_cpus() to prevent races in 3.4.

The partial traceback in previous comment is only a warning (not a panic) on at least some architectures, do you have a full traceback with a BUG/WARNING line?

It's not clear we need to cancel the timers on idle entry, maybe the timers were migrated to another CPU and should be canceled prior to a CPUFREQ_GOV_START that will re-add them on the proper CPU?

Patch-set: 3
Reviewer: Todd Poynor <1004277@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0