Update patch set 2
Patch Set 2:

The problem with getting this into stable is that the functionality only got into mainline in 3.11 or 3.12. This bug was discovered and fixed in ~3.12.4, so there didn't need to be any stable backports.

However, we use this functionality and have a CTS test for it, so we need to backport it to the branches we use.

What are the conflicts you're seeing? I remember that one vendor did their own backport of this in order to pass the CTS test, and instead of cherry-picking the upstream fixes as I expected them to, they did it all together in fudged patch. Is that what you're running into?

In hindsight, should this have been done differently?

Patch-set: 2
Reviewer: Lorenzo Colitti <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0