Update patch set 1
Patch Set 1:

https://android-review.googlesource.com/#/c/94372/ modifies keyreset.c to be on top of this driver, so there isn't a significant increase in in-tree code.

This driver will always require platform data, its purpose is to call a board-file defined callback function, which is something that can't be described in device tree.  Describing the delay to reset is also a little iffy in device tree - that's a policy, not a hardware description.  We could split the gpio configuration into device tree, but this is mostly refactoring, so that would be a separate patch.

Patch-set: 1