Update patch set 3
Patch Set 3:

> Was there any reason why SE_SBLABELSUPP (circa 3.10) or
 > selinux_is_sblabel_mnt() (circa 3.19) was not applied universally
 > instead of sysfs/rootfs only for GENFS? Just curious.

seclabel option in /proc/self/mounts is indicator to userspace that it can perform setxattr of security.selinux on files in that filesystem and safely expect it to work and be preserved for the lifetime of the file.  We originally did not prevent setxattr of files in filesystems labeled using genfs, but this was unreliable in general and required userspace to maintain its own list of filesystem types that were known to behave correctly, so seclabel mount option was introduced and made a requirement for setxattr of security.selinux.  So it is case-by-case for the genfs-labeled filesystems.

Patch-set: 3
Label: Verified=0