Update patch set 3
Patch Set 3:

Need your advice here.

if I make "android_usb" specific to android, whose kobj would I use in
kobject_uevent_env to send the uevent notification ? ("android0" device is the one we are using at present. 

Or is it just a case of renaming the sysfs class and the device in a more generic way ?

Patch-set: 3
Label: Verified=0