Update patch set 8
Patch Set 8:

Should have some actual numbers by the end of the week. We are anticipating that the impact will be very low because 1) every ioctl is already hooked by selinux and a permission check is performed - so we get that for free. 2) the ioctl command decision is cached along with the ioctl permission decision, so only the existing cache lookup occurs. 3) the additional check on the ioctl command will be optional, will only be applied to a select set of ioctls that are accessible to untrusted_app, and should we decide that the overhead is too great on that particular device e.g. gpu, we can revert to the previous permission-only check. I'm sure some actual data will be much more reassuring, so I'll get those out asap. Let me know if there are any benchmarks that you would like me to run.

Patch-set: 8