Update patch set 11
Patch Set 11:

Wondering if I shouldn't start a new patch series as this is a complete redo of the architecture. This patch set addresses limitations on the number of ioctl commands that can be input. It also improves lookup time for command permissions resulting in a performance improvement.

Patch-set: 11
Label: Verified=0