Update patch set 1
Patch Set 1:

Hi Rahul, I agree with the patch in general that load should be calculated based on the actual frequency CPU is running at. We have internal patches doing same thing as well, except the re-arm part I pointed out because we can't risk losing the slack timer while target_freq could be non-min.

I plan to upload a few patches that qcom uses for comments soon, but you beat me to it for this change. :-)

Patch-set: 1