Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

The only benefit of not running the interactive timer when running at max_freq is the time gained by avoiding the execution of the cpufreq_interactive_timer callback.(we know that we are going to stay at max_freq if the cpu is still sufficiently busy) Since, this is exactly intended for the max_freq, I wasn't sure of how this would affect the benchmarks(any small time gained during cpu_benchmarks is always good) across platforms and hence only focused on NOT rescheduling the timers when policy->min is changed.                                   

Given the inconsistencies stated, yes it does seem to be good to always have the timers running.

if the target_freq gets constrained by the change in the policy->max in the middle of the timer interval, are we relying on the cpufreq_notifier to update the load(update_load) ?

Patch-set: 3
Label: Code-Review=+1