Update patch set 4
Patch Set 4:

You are right about FW path and this is my shame. It is not generic. We want to add LTECOEX support not only for bcm4358, but for the platform. This is why I am asking what userspace part is going to use it.

Patch-set: 4
Reviewer: Dmitry Shmidt <1000414@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0