Update patch set 2
Patch Set 2:

(c) Kevin:
According to my understanding:
If the thread is running when the event is enqueued, then we'll have +1 queued event, +1 wakelock, +1 semaphore count
If the thread is not running when the event is enqueued, then we'll have +1 queued event, +0 wakelock, +0 semaphore count (wl_wakeup_event() becomes a no-op because thr_pid is -1 before PROC_START or after PROC_STOP)
In the "not running" case we'll want to dequeue the event but we won't want to release a wakelock: -1 queued events, +0 wakelock, +0 semaphore count.  The way to detect this case is to match down() calls with DHD_OS_WAKE_UNLOCK() calls.

Patch-set: 2