Update patch set 2
Patch Set 2:

(2 comments)

This change is needed in order to enable addition of a hook in IPCThreadState.cpp to gate dispatch of transactions to known binder services based on service descriptor and transaction code.  getpidcon() is insufficient since pid is not passed along with one-way binder transactions, due to possibility of process death.

Patch-set: 2