Update patch set 3
Patch Set 3:

It has been a week since submitted upstream, deathly silence in response. IMHO this happens when the vast majority of the developers are focussed on stabilizing a release. Or they think OMG, you are putting test code inside the kernel, WTH are you thinking?

I will consider any chance of upstreaming this abandoned for the moment (but may ping again in a month). Looking for +2 now, and permission to also propagate this to 3.10, 3.14, 3.18 and 4.4 Android common kernels.

Patch-set: 3