Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

I expect this is needed in 3.10, 3.14, 3.18, 4.1 and 4.4?

Only _very_ minor nits:

If from upstream or list, leave as is. But would have liked to have seen this as two commits, one for the movement of the define, and another for the Module export statements?

Patch-set: 1
Label: Code-Review=+1