Update patch set 1
Patch Set 1:

> (2 comments)
 > 
 > I expect this is needed in 3.10, 3.14, 3.18, 4.1 and 4.4?

As I commented above (but noticed after submitting), this is already in android-3.14. It's been kept in chromeos-3.18, but it never made it to android-3.18. I didn't check the others yet.

 > Only _very_ minor nits:
 > 
 > If from upstream or list, leave as is. But would have liked to have
 > seen this as two commits, one for the movement of the define, and
 > another for the Module export statements?

I'd prefer leaving it as-is, since it's in several repos as-is already.

I can reword any commit messages to clarify anything you'd like. For instance, I'll add that it was cherry-picked from android-3.14. And unless you object, I guess I'll add the "ANDROID:" prefix, since these tracepoints are only in android (and android-derivative, like chromium) kernels.

Patch-set: 1