Update patch set 1
Patch Set 1:

Is there a status update ?

Maybe there is a better solution, but for my part I don't understand the file system code well enough to propose one. For the time being, it would be great if someone with +2 privilege can approve this patch to fix the build failure.

Patch-set: 1