Update patch set 4
Patch Set 4:

Submit button is present, you just can not see it because your permissions are not high enough ;-}. As itemized in my proposed workflow for getting CLs upstreamed in kernel.org, the next step is to cherry-pick or BACKPORT: this change to the other active android (and if relevant, chrome) kernels (hint: tie them together with Topic so they can be submitted in unison once all BACKPORT's are cleared). Cherry pick / BACKPORT to 3.10, 3.14, 3.18 & 4.1. Final clearance will require verification that your change compiles _and_ works on a clean merge-down on one of the actively developed devices. At that point one of us in the kernel group would be pleased to hit the Submit button for you (or grant you the permissions to do so)

Patch-set: 4
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0