Update patch set 3
Patch Set 3:

Not sure yet where is the best place to handle the trap in a generic way.
All arch are calling notify_die generic function.
We could add a line there, not great, we could have RTC register a die handler but rtc-lib is not a device in any way , so not great either ...

Patch-set: 3
Reviewer: Thierry Strudel <1071633@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0