Update patch set 1
Patch Set 1:

This is an Android N specific CL due to the reduced permission on sysfs that has been enforced in N.
I don't think it make sense to upstream this specific CL as what would make sense to upstream is a broader cleanup of the uevents generation in case of CPU hotplug.
Thanks to cherry-pick to the supported android common kernel branches.

Patch-set: 1
Reviewer: Thierry Strudel <1071633@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0