EXP: sched/fair: use boosted task utilisation to constraint CPU selection
A boosted task needs to be scheduled on a CPU which can grant a minimum
capacity which is higher than its utilization.
However, a task can also be allocated on a CPU which already provides a
utilization higher than the boosted task's utilization itself.

With the current approach, the more a task is boosted the more likely is to
be considered not fitting on non idle CPUs.  For example, a task 100%
boosted is not fitting any non idle CPU and thus we always schedule it on
the previous CPU.

This patch uses the boosted task utilization as a threashold (min_util)
which defines the minimum capacity that must be available on a CPU to be
a valid next_cpu for that task.

NOTE: this patch in re-introducing the same policy used before:
   0fe9a8cb2 sched/fair: Using boosted_task_util to calculate new_util in energy_aware_wake_cpu
which is claiming to improves performance of Vellamo suite by ~8% to 10%.
However, this patch has been followed by:
   a09d6f365 sched: eas: optimize idle cpu selection for boosted tasks
where we introduces the policy to alwasy pick an IDLE CPU for boosted tasks.

With this last patch we can avoid using the boosted_task_utilization just
to favour moving a boosted task into an IDLE CPU, which was quite likely
the ultimate effect of 0fe9a8cb2.

Change-Id: I194492b9e9b8fb35bab4c26095c218146903dff4
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Srinath Sridharan <srinathsr@google.com>