Update patch set 5
Patch Set 5:

I just hit the 'submit feedback' button on the page and asked them to update to latest android-common kernel. I figure that will get to the right documentation person to be tracked ;-}

There is an issue in that 4.1 and 4.4 cleared with SECCOMP_FILTER=y, please submit an update CL to those that drop that entry as not required so that the Kernel Configuration page is correct.

Patch-set: 5