Update patch set 1
Patch Set 1: Code-Review-1

I need a better rationalization, are the other upstream fixes necessary to make this change relevant or not? In the current codebase, it may be advised to move the NULLing of private before the call to the completion handler instead? Letting badhri@ decide as he knows this code better.

Patch-set: 1
Label: Code-Review=-1