Update patch set 2
Patch Set 2:

(3 comments)

There is a lot of duplicate checks added by this change. Can you think of a way to reduce this? It is also a little odd to separate out fd objects, but keep the handle and ptr types as using members using a shared struct.

Patch-set: 2
Label: Verified=0