sched/trace: Add task migration cause to trace_sched_migrate_task
Enhance the trace event 'sched_migrate_task' for CFS tasks by providing
the actual cause of the task migration.

This is useful in case you want to evaluate the migration history of a
specific task of a workload.

This feature can be enabled by setting CONFIG_SCHED_DEBUG_EAS_MIGRATION
(disabled by default).

Change-Id: I8c2e58fac083b39424d03b9ae24dcddd7d28ae71
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>