sched/trace: Add task migration cause to trace_sched_migrate_task
Enhance the trace event 'sched_migrate_task' for CFS tasks by providing
the actual cause of the task migration.

This is useful in case you want to evaluate the migration history of a
specific task of a workload.

This feature can be enabled by setting CONFIG_SCHED_DEBUG_EAS_MIGRATION
(disabled by default).

Change-Id: I2c8082d86a6c6ad5f3880b431a1239e3dd7a5333
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>