android: binder: hwbinder devnode.
In the Treble model, we'll be using the kernel
binder driver not only for regular platform IPC,
but also for IPC to HAL implementations, which
will live in their own process in the Treble world.

Since we want to cleanly separate platform IPC
from so-called "vendor IPC", we want to use separate
context managers for the two domains. This is
achieved by mapping the context manager to a
device node, and introducing a new device node
(/dev/hwbinder) for the vendor domain.

Change-Id: I3df72b2a19b5ad5a0360e6322482db7b00a12b24
Signed-off-by: Martijn Coenen <maco@google.com>