Update patch set 5
Patch Set 5:

The existing selinux checks do not only apply to service lookup. There are hooks the allow you to restrict which processes can call into each other and if they can transfer objects (although not a specific set of objects) and specific files. I would think it would be useful for these rules to be different for each binder context.

Patch-set: 5