Update patch set 13
Patch Set 13:

> Another meta-comment, why not make this all dynamic?  Have
 > libbinder create the new instances when needed, how are you going
 > to know the exact number of instances at build time?

We currently anticipate only two binder "domains", and expect this to be pretty static. I made it more generic because there was talk of using a separate binder domain for core graphics processes, but that looks unlikely now (fine-grained locking work is in progress). Even so, that would have been known that at compile-time for a device.

Out of interest, if we were to make this dynamic, an ioctl() seems like the only way to go?

As for the userspace story, I'll point you to a design doc.

 > 
 > This is overflow from the Ara work where everything was done
 > dynamic on the fly (adding new cameras, audio devices, displays,
 > etc.)  Don't bake into the new api static build time dependencies
 > if there is any chance to make it dynamic.

Patch-set: 13
Label: Verified=0