Update patch set 4
Patch Set 4:

Moving the debugfs files to the context affects userspace (your hwbinder debug information will not be part of a bugreport unless you make a matching change in user-space). If we want that change it should be separate from a change that moves internal state around to avoid lock contention while we wait for a real fix to that issue.

Patch-set: 4