Update patch set 4
Patch Set 4:

I'm not saying it makes the code less readable. If you want to change a driver that other people also work on, your change has be an improvement, not just different.
I want proof that this change has a noticeable improvement before merging it. What is the overhead of moving HALs to use hwbinder with and without this patch. Your example is just a specific case of a problem that is being actively worked on. Why would a hwbinder transaction and binder transaction be more likely to need the lock at the same time than two hwbinder transactions or two binder transactions?

Patch-set: 4