Update patch set 23
Patch Set 23:

I have not finished reviewing the latest version yet. One question I have not seen answer to is what happens if these new object types are passed to libbinder. That cannot happen without this change, so we need to make sure libbbinder does not blow up in that case.

Patch-set: 23