Update patch set 4
Patch Set 4:

> I'm not saying it makes the code less readable. If you want to
 > change a driver that other people also work on, your change has be
 > an improvement, not just different.
 > I want proof that this change has a noticeable improvement before
 > merging it. What is the overhead of moving HALs to use hwbinder
 > with and without this patch. Your example is just a specific case
 > of a problem that is being actively worked on. Why would a hwbinder
 > transaction and binder transaction be more likely to need the lock
 > at the same time than two hwbinder transactions or two binder
 > transactions?

Two binder or two hwbinder transactions are still bad with this change - I think we all agree the fine-grained locking approach is the better way to go, eventually. The point is that we are adding a *lot* of traffic that used to be simple in-process function calls to binder IPC now; so there will be a lot more contention for the same lock. So in my mind there is no doubt that this change will be an improvement. It will certainly improve even more by the time we have fine-grained locking, but that change has not landed yet, and in the mean time we are trying to do what we can.

Patch-set: 4
Label: Verified=0