trusty: add support for SM Wall object
SM Wall is a shared memory buffer established between secure and
non-secure side that allows for secure side to publish in efficient
manner certain state that non-secure side might acts.

This patch adds support for such buffer in a generic way, an API to
setup such buffer with secure side and an API to locate it's content
based on well object known id's.

Change-Id: I66bf79190e18635b7df6800385f4284c1095b786
Signed-off-by: Michael Ryleev <gmar@google.com>