trusty: add support for trusty backup timer
On some platforms, in certain cpu idle modes, Trusty might
lose the state of secure timer that it is using for work scheduling.
In such cases, non-secure side would typically migrate such timers to
alternative implementations that does not lose their state. Ideally,
secure side should have similar mechanizm, but it might not be always
feasible due to hardware limitations.

This patch introduces a generic workaround for this issue but adding
backup non-secure timers that is used to kick cpus out of deep idle
modes when appropriate.

Change-Id: I6712aaea7d6c34ee5e9018f44888bb2428380d11
Signed-off-by: Michael Ryleev <gmar@google.com>