Update patch set 1
Patch Set 1:

In virtualization TEE implementation,  as we don't want to virtualize a virtual secure timer for Trusty due to performance reason, can we use this Linux "backup timer" for Trusty scheduling? also we're enabling trusty dynamic timer.

For example, if trusty app calls a sleep function in trusty (e.g. 30s), and then if there is nothing to do in Trusty, the VMM will schedule to non-secure world... and when 30s expires in non-secure world, there should have an timer fire event happening in linux trusty driver, then trusty driver call an API to switch to Trusty world.

can this feature meet it?

Patch-set: 1