Update patch set 1
Patch Set 1:

This one is not intended to replace secure timer, just to provide a generic workaround for some artifacts we have seen on some platforms. This solution is not ideal, it would be better to implement power persistent timer using real hardware block but it is impossible to to do in a generic way. There are two primary usages for secure timer in Trusty: it is used as preemption timer (which is not really working very well due to some artifacts in Lk scheduling and it's interaction with non-secure side) and as a source for all app timers which is app specific. If you choose not to properly implement secure timer (fake it) the most apps would probably run, but latency of app timers will be very inconsistent.

Patch-set: 1
Reviewer: Michael Ryleev <1045553@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0