Update patch set 1
Patch Set 1:

Hi,

I'm not sure I get where the problem is looking at the backtrace.
It looks like update_min_max_capacity() acquires lot of nested rq->lock and this confuses LOCKDEP?

I was able to reproduce the issue on an hikey board and the following seems to cure the warning.

 static void update_min_max_capacity(void)
 {
        unsigned long flags;
-       int i;
+       int i, idx = 0;
 
        local_irq_save(flags);
-       for_each_possible_cpu(i)
-               raw_spin_lock(&cpu_rq(i)->lock);
+       for_each_possible_cpu(i) {
+               if (!idx)
+                       raw_spin_lock(&cpu_rq(i)->lock);
+               else
+                       raw_spin_lock_nested(&cpu_rq(i)->lock, idx);
+               idx++;
+       }
 
        __update_min_max_capacity();

Patch-set: 1
Label: Verified=0