sched/rt: fix rt_nr_cfs_throttled inconsistency after migrations
Removes rt_nr_cfs_throttled and instead uses emptiness of
cfs_throttled_tasks to findout if there are demoted tasks for a given
rt_rq.  This way after demoted tasks migrate, there is no mismatch
bewteen the count which was updated on the potentially wrong rt_rq
and the size of the list.

Signed-off-by: Andres Oportus <andresoportus@google.com>
Change-Id: I76bb629da161a092401e7b278a26eddf1aab0676