Update patch set 4
Abandoned

I included this commit in the change stack only because we need https://android-review.googlesource.com/#/c/312353/4 ("tcp: fix more NULL deref after prequeue changes"), and that commit was written as a fix for this one.

However, this commit is not a dependency, and it's not clear whether it's safe to backport it to 3.10, whether it needs 093162553c33 ("tcp: force a dst refcount when prequeue packet") to be correct, etc.

Because https://android-review.googlesource.com/#/c/312353/4 seems correct and itself, and does not depend on this one, it seems better to skip this one.

Patch-set: 4
Status: abandoned
Tag: autogenerated:gerrit:abandon