Update patch set 6
Patch Set 6:

Sorry about the delay -- we haven't ignored this. There is some question about whether step #5 is a supported operation (sending sync transaction during BBinder destructor) which we want to talk to someone in our frameworks team. I suspect with the holidays, we won't have that conversation until January, so this will have to wait until then.

Generally, it seems like using the proc todo is probably ok. I confirmed that our tests pass with this change.

If we do go ahead with this, it will need to be resubmitted against android-4.4 (not android-4.4.y).

Patch-set: 6
Label: Verified=0