sched/rt: fix demotion paths when switching from userspace
Demotion may be triggered in any update_curr_rt() invocation
which may happen when dequeing an RT task being switched to
CFS from userspace.  This leaves the task being switched in
demoted CFS and not deuqued as expected by __sched_setcheduler()
in the middle of its class switch.
Demotion may also be done in __enqueue_rt_task() if the
rt_rq is throttled but the rt_se is not, this intends to catch
the wakeup scenario but may be called from a userspace switch
for instance when enqueuing an RT task.

Signed-off-by: Andres Oportus <andresoportus@google.com>